\name{dizzysNewInfec-package}

\alias{dizzysNewInfec-package}

\alias{dizzysNewInfec}

\docType{package}

\title{Modelling of Infectious Diseases}

\description{Permiting to simulate infectious diseases by using the SIR/SEIR models under deterministic and stochastic models.}

\details{

\tabular{ll}{
Package: \tab dizzysNewInfec\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2013-05-02\cr
License: \tab R \cr
Depends: \tab methods, deSolve, rgl, survival, KMsurv\cr
}
}

\author{TRAN Thi Cam Giang <camgiang2010@gmail.com>}

\references{Matt J. Keeling and Pejman Rohani (2008) Modeling Infectious Diseases IN HUMANS AND ANIMALS. Princeton University Press.

	Cao Y, Gillespie DT, Petzold LR,The Journal of Chemical Physics, 2007.

	Norman Matloff (2009), The Art of R Programming.}

\keyword{ package }

\seealso{package 'dizzyz'}
